fix chi square calculation

+++ src/main/java/org/example/CustomPair.java
@@ -2,22 +2,22 @@ package org.example;
 
 public class CustomPair implements Comparable<CustomPair> {
     private String key;
-    private float value;
-    public CustomPair(String key, float value) {
+    private double value;
+    public CustomPair(String key, double value) {
         this.key = key;
         this.value = value;
     }
 
     @Override
     public int compareTo(CustomPair o) {
-        return Float.compare(this.value, o.value);
+        return Double.compare(this.value, o.value);
     }
 
     public String getKey() {
         return key;
     }
 
-    public float getValue() {
+    public double getValue() {
         return value;
     }
 }
